+++ test/extensions/access_loggers/grpc/BUILD
@@ -27,6 +27,18 @@ envoy_extension_cc_test(
 )
 
 envoy_extension_cc_test(
+    name = "grpc_access_log_utils_test",
+    srcs = ["grpc_access_log_utils_test.cc"],
+    extension_name = "envoy.access_loggers.http_grpc",
+    deps = [
+        "//source/extensions/access_loggers/grpc:grpc_access_log_utils",
+        "//test/mocks/local_info:local_info_mocks",
+        "//test/mocks/ssl:ssl_mocks",
+        "//test/mocks/stream_info:stream_info_mocks",
+    ],
+)
+
+envoy_extension_cc_test(
     name = "http_grpc_access_log_impl_test",
     srcs = ["http_grpc_access_log_impl_test.cc"],
     extension_name = "envoy.access_loggers.http_grpc",
